requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas
numpy
python-dotenv
langchain-community
fastapi
uvicorn
langchain-openai
sqlalchemy
psycopg2-binary
langchain_core
pinecone
transformers
langchain_community
sentence-transformers
